Move UNIT_CREATE to cUnits
