+++ src/gameobjects/structures/cOrderProcesser.cpp
@@ -238,7 +238,7 @@ void cOrderProcesser::sendFrigate()
         }
         else {
             // STEP 2: create frigate
-            int unitId = UNIT_CREATE(iStartCell, FRIGATE, player->getId(), true);
+            int unitId = cUnits::UNIT_CREATE(iStartCell, FRIGATE, player->getId(), true);
             if (unitId < 0) {
                 logbook("cOrderProcesser::sendFrigate : unable to spawn frigate");
                 return;
