+++ src/gameobjects/units/cUnits.h
@@ -85,6 +85,23 @@ public:
     auto end() { return m_units.end(); }
     auto end() const { return m_units.end(); }
 
+    /**
+    * creates a unit, the isReinforcement flag is true when the unit is created for / by reinforcements. This
+    * flag will make sure to trigger a different event type (not CREATED, but REINFORCED) so that we can distinguish
+    * between them.
+    *
+    * @param iCll
+    * @param unitType
+    * @param iPlayer
+    * @param bOnStart
+    * @param isReinforcement
+    * @return
+    */
+    static int UNIT_CREATE(int iCll, int unitType, int iPlayer, bool bOnStart);
+    static int UNIT_CREATE(int iCll, int unitType, int iPlayer, bool bOnStart, bool isReinforcement);
+    static int UNIT_CREATE(int iCll, int unitType, int iPlayer, bool bOnStart, bool isReinforcement, float hpPercentage);
+
+
 private:
     // Maximum number of units
     static constexpr int MAX_UNITS_CAPACITY = 300;
